refactor(os/interfaces): rename shared.py → common.py
"shared" is too generic — it doesn't say WHAT is shared or WITH WHOM.
"common" is conventional Python naming for cross-module utilities and
reads naturally at import sites:

  from agno.os.interfaces.common import build_session_store_config

Sits alongside base.py without an underscore-prefix (they're both first-
class public modules). Scope unchanged — same helpers, same signatures.

Updated 5 import sites: 4 Discord files (events.py, security.py,
router.py, state.py) + 1 WhatsApp file (router.py).
